odbms/database.py
from typing import Dict, List, Any, Optional
import asyncio

class Database:
    """Base class for database implementations."""
    
    def __init__(self, **kwargs):
        """Initialize database connection."""
        self.dbms = kwargs.get('dbms', None)
        self.connection = None
        self.config = kwargs
    
    def connect(self) -> None:
        """Connect to the database."""
        raise NotImplementedError
    
    def disconnect(self) -> None:
        """Disconnect from the database."""
        raise NotImplementedError
    
    def execute(self, query: str, params: Optional[Dict[str, Any]] = None) -> Any:
        """Execute a raw query."""
        raise NotImplementedError
    
    def find(self, table: str, conditions: Optional[Dict[str, Any]] = None) -> List[Dict[str, Any]]:
        """Find records matching conditions."""
        raise NotImplementedError
    
    def find_one(self, table: str, conditions: Dict[str, Any]) -> Optional[Dict[str, Any]]:
        """Find a single record matching conditions."""
        raise NotImplementedError
    
    def insert(self, table: str, data: Dict[str, Any]) -> Any:
        """Insert a record."""
        raise NotImplementedError
    
    def insert_many(self, table: str, data: List[Dict[str, Any]]) -> int:
        """Insert multiple records."""
        raise NotImplementedError
    
    def update(self, table: str, conditions: Dict[str, Any], data: Dict[str, Any]) -> int:
        """Update records matching conditions."""
        raise NotImplementedError
    
    def remove(self, table: str, conditions: Dict[str, Any]) -> int:
        """Remove records matching conditions."""
        raise NotImplementedError
    
    async def execute_async(self, query: str, params: Optional[Dict[str, Any]] = None) -> Any:
        """Execute a raw query asynchronously."""
        return await asyncio.get_event_loop().run_in_executor(
            None, self.execute, query, params
        )
    
    async def find_async(self, table: str, conditions: Optional[Dict[str, Any]] = None) -> List[Dict[str, Any]]:
        """Find records matching conditions asynchronously."""
        return await asyncio.get_event_loop().run_in_executor(
            None, self.find, table, conditions
        )
    
    async def find_one_async(self, table: str, conditions: Dict[str, Any]) -> Optional[Dict[str, Any]]:
        """Find a single record matching conditions asynchronously."""
        return await asyncio.get_event_loop().run_in_executor(
            None, self.find_one, table, conditions
        )
    
    async def insert_async(self, table: str, data: Dict[str, Any]) -> Any:
        """Insert a record asynchronously."""
        return await asyncio.get_event_loop().run_in_executor(
            None, self.insert, table, data
        )
    
    async def insert_many_async(self, table: str, data: List[Dict[str, Any]]) -> int:
        """Insert multiple records asynchronously."""
        return await asyncio.get_event_loop().run_in_executor(
            None, self.insert_many, table, data
        )
    
    async def update_async(self, table: str, conditions: Dict[str, Any], data: Dict[str, Any]) -> int:
        """Update records matching conditions asynchronously."""
        return await asyncio.get_event_loop().run_in_executor(
            None, self.update, table, conditions, data
        )
    
    async def remove_async(self, table: str, conditions: Dict[str, Any]) -> int:
        """Remove records matching conditions asynchronously."""
        return await asyncio.get_event_loop().run_in_executor(
            None, self.remove, table, conditions
        )
